Use global EndpointConfig
