+++ rasa/nlu/data_router.py
@@ -1,3 +1,4 @@
+import asyncio
 import datetime
 import io
 import logging
@@ -102,7 +103,8 @@ class DataRouter(object):
         else:
             self.component_builder = ComponentBuilder(use_cache=True)
 
-        self.project_store = self._create_project_store(project_dir)
+        self.project_store = asyncio.run(self._create_project_store(
+            project_dir))
 
         # tensorflow sessions are not fork-safe,
         # and training processes have to be spawned instead of forked. See
@@ -156,7 +158,7 @@ class DataRouter(object):
         projects.extend(self._list_projects_in_cloud())
         return projects
 
-    def _create_project_store(self,
+    async def _create_project_store(self,
                               project_dir: Text) -> Dict[Text, Any]:
         default_project = RasaNLUModelConfig.DEFAULT_PROJECT_NAME
 
@@ -165,7 +167,7 @@ class DataRouter(object):
         project_store = {}
 
         if self.model_server is not None:
-            project_store[default_project] = load_from_server(
+            project_store[default_project] = await load_from_server(
                 self.component_builder,
                 default_project,
                 self.project_dir,
