+++ rasa/nlu/project.py
@@ -30,7 +30,7 @@ STATUS_TRAINING = 1
 STATUS_FAILED = -1
 
 
-def load_from_server(component_builder: Optional[ComponentBuilder] = None,
+async def load_from_server(component_builder: Optional[ComponentBuilder] = None,
                      project: Optional[Text] = None,
                      project_dir: Optional[Text] = None,
                      remote_storage: Optional[Text] = None,
@@ -45,7 +45,7 @@ def load_from_server(component_builder: Optional[ComponentBuilder] = None,
                       remote_storage=remote_storage,
                       pull_models=True)
 
-    _update_model_from_server(model_server, project)
+    await _update_model_from_server(model_server, project)
 
     if wait_time_between_pulls:
         # continuously pull the model every `wait_time_between_pulls` seconds
@@ -55,7 +55,7 @@ def load_from_server(component_builder: Optional[ComponentBuilder] = None,
     return project
 
 
-def _update_model_from_server(model_server: EndpointConfig,
+async def _update_model_from_server(model_server: EndpointConfig,
                               project: 'Project') -> None:
     """Load a zipped Rasa NLU model from a URL and update the passed
 
@@ -65,7 +65,7 @@ def _update_model_from_server(model_server: EndpointConfig,
 
     model_directory = tempfile.mkdtemp()
 
-    new_model_fingerprint, filename = _pull_model_and_fingerprint(
+    new_model_fingerprint, filename = await _pull_model_and_fingerprint(
         model_server, model_directory, project.fingerprint)
     if new_model_fingerprint:
         model_name = _get_remote_model_name(filename)
@@ -87,7 +87,7 @@ def _get_remote_model_name(filename: Optional[Text]) -> Text:
         return MODEL_NAME_PREFIX + timestamp
 
 
-def _pull_model_and_fingerprint(model_server: EndpointConfig,
+async def _pull_model_and_fingerprint(model_server: EndpointConfig,
                                 model_directory: Text,
                                 fingerprint: Optional[Text]
                                 ) -> (Optional[Text], Optional[Text]):
@@ -99,7 +99,7 @@ def _pull_model_and_fingerprint(model_server: EndpointConfig,
     try:
         logger.debug("Requesting model from server {}..."
                      "".format(model_server.url))
-        response = model_server.request(method="GET",
+        response = await model_server.request(method="GET",
                                         headers=header,
                                         timeout=DEFAULT_REQUEST_TIMEOUT)
     except RequestException as e:
@@ -133,11 +133,11 @@ def _pull_model_and_fingerprint(model_server: EndpointConfig,
     return response.headers.get("ETag"), response.headers.get("filename")
 
 
-def _run_model_pulling_worker(model_server: EndpointConfig,
+async def _run_model_pulling_worker(model_server: EndpointConfig,
                               wait_time_between_pulls: int,
                               project: 'Project') -> None:
     while True:
-        _update_model_from_server(model_server, project)
+        await _update_model_from_server(model_server, project)
         time.sleep(wait_time_between_pulls)
