+++ tests/core/test_actions.py
@@ -12,7 +12,7 @@ from rasa.core.actions.action import (
 from rasa.core.domain import Domain
 from rasa.core.events import Restarted, SlotSet, UserUtteranceReverted
 from rasa.core.trackers import DialogueStateTracker
-from rasa.core.utils import ClientResponseError, EndpointConfig
+from rasa.utils import ClientResponseError, EndpointConfig
 from tests.core.utilities import json_of_latest_request, latest_request
