+++ tests/core/test_utils.py
@@ -105,51 +105,6 @@ def test_read_lines():
     assert len(lines) == 2
 
 
-async def test_endpoint_config():
-    with aioresponses() as mocked:
-        endpoint = EndpointConfig(
-            "https://example.com/",
-            params={"A": "B"},
-            headers={"X-Powered-By": "Rasa"},
-            basic_auth={"username": "user",
-                        "password": "pass"},
-            token="mytoken",
-            token_name="letoken",
-            type="redis",
-            port=6379,
-            db=0,
-            password="password",
-            timeout=30000
-        )
-
-        mocked.post('https://example.com/test?A=B&P=1&letoken=mytoken',
-                    payload={"ok": True},
-                    repeat=True,
-                    status=200)
-
-        await endpoint.request("post", subpath="test",
-                               content_type="application/text",
-                               json={"c": "d"},
-                               params={"P": "1"})
-
-        r = latest_request(mocked, 'post',
-                           "https://example.com/test?A=B&P=1&letoken=mytoken")
-
-        assert r
-
-        assert json_of_latest_request(r) == {"c": "d"}
-        assert r[-1].kwargs.get("params", {}).get("A") == "B"
-        assert r[-1].kwargs.get("params", {}).get("P") == "1"
-        assert r[-1].kwargs.get("params", {}).get("letoken") == "mytoken"
-
-        # unfortunately, the mock library won't report any headers stored on
-        # the session object, so we need to verify them separately
-        async with endpoint.session() as s:
-            assert s._default_headers.get("X-Powered-By") == "Rasa"
-            assert s._default_auth.login == "user"
-            assert s._default_auth.password == "pass"
-
-
 os.environ['USER_NAME'] = 'user'
 os.environ['PASS'] = 'pass'
