+++ tests/utilities.py
@@ -1,0 +1,8 @@
+from yarl import URL
+
+
+def latest_request(mocked, request_type, path):
+    return mocked.requests.get((request_type, URL(path)))
+
+def json_of_latest_request(r):
+    return r[-1].kwargs["json"]
